+++ jvm/src/main/scala/gpdviz/async/PusherNotifier.scala
@@ -1,6 +1,7 @@
 package gpdviz.async
 
 import com.pusher.rest.Pusher
+import gpdviz._
 import gpdviz.config.{PusherCfg, cfg}
 import gpdviz.model.{DataStream, ObsData, SensorSystem}
 import gpdviz.server.JsonImplicits
@@ -34,6 +35,13 @@ class PusherNotifier(pusherCfg: PusherCfg) extends Notifier with JsonImplicits{
       ss.description foreach { v ⇒ map = map + ("description" → v) }
       ss.clickListener foreach { v ⇒ map = map + ("clickListener" → v) }
       notifyEvent(ss.sysid, "sensorSystemRegistered", map.asJava)
+
+      notifyEvent2(ss.sysid, SensorSystemRegistered(
+        sysid = ss.sysid,
+        name = ss.name,
+        description = ss.description,
+        clickListener = ss.clickListener
+      ))
     }
   }
 
@@ -44,6 +52,11 @@ class PusherNotifier(pusherCfg: PusherCfg) extends Notifier with JsonImplicits{
         "str" -> str.toJson.compactPrint
       )
       notifyEvent(ss.sysid, "streamAdded", map.asJava)
+
+      notifyEvent2(ss.sysid, StreamAdded(
+        sysid = ss.sysid,
+        str = str.toJson.compactPrint
+      ))
     }
   }
 
@@ -61,6 +74,7 @@ class PusherNotifier(pusherCfg: PusherCfg) extends Notifier with JsonImplicits{
       }
       map.asJava
     }
+
     @tailrec
     def rec(from: Int): Unit = {
       if (from < obs.size) {
@@ -72,17 +86,52 @@ class PusherNotifier(pusherCfg: PusherCfg) extends Notifier with JsonImplicits{
       }
     }
     rec(0)
+
+    notifyObservations2AddedNEW(ss, strid, observations)
+  }
+
+  private def notifyObservations2AddedNEW(ss: SensorSystem, strid: String, observations: Map[String, List[ObsData]]): Unit = if (ss.pushEvents) {
+    val obs = observations mapValues { list ⇒
+      val obsDataSTRs = collection.mutable.ListBuffer[ObsDataSTR]()
+      list foreach  { o ⇒
+        obsDataSTRs += ObsDataSTR(
+          feature = o.feature.map(_.toJson.compactPrint),
+          geometry = o.geometry.map(_.toJson.compactPrint),
+          scalarData = o.scalarData
+        )
+      }
+      obsDataSTRs.toList
+    }
+
+    @tailrec
+    def rec(from: Int): Unit = {
+      if (from < obs.size) {
+        val next = Math.min(from + 15, obs.size)
+        val slice = obs.slice(from, next)
+        notifyEvent2(ss.sysid, Observations2Added(
+          sysid = ss.sysid,
+          strid = strid,
+          obss = slice
+        ))
+        rec(next)
+      }
+    }
+    rec(0)
   }
 
   def notifyStreamRemoved(ss: SensorSystem, strid: String): Unit = {
     val map = Map("sysid" -> ss.sysid, "strid" -> strid)
     notifyEvent(ss.sysid, "streamRemoved", map.asJava)
+
+    notifyEvent2(ss.sysid, StreamRemoved(ss.sysid, strid))
   }
 
   def notifySensorSystemUpdated(ss: SensorSystem): Unit = {
     if (ss.pushEvents) {
       val map = Map("sysid" -> ss.sysid)
       notifyEvent(ss.sysid, "sensorSystemUpdated", map.asJava)
+
+      notifyEvent2(ss.sysid, SensorSystemUpdated(ss.sysid))
     }
   }
 
@@ -90,6 +139,8 @@ class PusherNotifier(pusherCfg: PusherCfg) extends Notifier with JsonImplicits{
     if (ss.pushEvents) {
       val map = Map("sysid" -> ss.sysid)
       notifyEvent(ss.sysid, "sensorSystemRefresh", map.asJava)
+
+      notifyEvent2(ss.sysid, SensorSystemRefresh(ss.sysid))
     }
   }
 
@@ -97,6 +148,8 @@ class PusherNotifier(pusherCfg: PusherCfg) extends Notifier with JsonImplicits{
     if (ss.pushEvents) {
       val map = Map("sysid" -> ss.sysid)
       notifyEvent(ss.sysid, "sensorSystemUnregistered", map.asJava)
+
+      notifyEvent2(ss.sysid, SensorSystemUnregistered(ss.sysid))
     }
   }
 
@@ -107,6 +160,14 @@ class PusherNotifier(pusherCfg: PusherCfg) extends Notifier with JsonImplicits{
     val res = pusher.trigger(channel, "my_event", map.asJava)
     if (res.getHttpStatus != 200)
       println(s"!!!notifyEvent: pusher.trigger ERROR: status=${res.getHttpStatus} message=${res.getMessage}")
+  }
+
+  private def notifyEvent2(sysid: String, notif: Notif): Unit = {
+    val channel = s"${cfg.serverName}-$sysid-2"
+    val res = pusher.trigger(channel, "my_event", upickle.default.write(notif))
+    if (res.getHttpStatus != 200)
+      println(s"!!!notifyEvent2: pusher.trigger ERROR: status=${res.getHttpStatus} message=${res.getMessage}")
+
   }
 
   private val pusher = new Pusher(pusherCfg.appId, pusherCfg.key, pusherCfg.secret)