+++ shared/src/main/scala/gpdviz/model/model.scala
@@ -1,0 +1,13 @@
+package gpdviz.model
+
+case class SensorSystemSummary(sysid:        String,
+                               name:         Option[String] = None,
+                               description:  Option[String] = None,
+                               streamIds:    Set[String] = Set.empty
+                              )
+
+case class LatLon(lat: Double, lon: Double)
+
+case class ScalarData(vars:      List[String],
+                      vals:      List[Double],
+                      position:  Option[LatLon] = None)